% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.super_riesz}
\alias{predict.super_riesz}
\title{Predict from SuperRiesz object}
\usage{
\method{predict}{super_riesz}(object, newdata, ...)
}
\arguments{
\item{object}{SuperRiesz object}

\item{newdata}{data to use for predictions}

\item{...}{additional arguments (not currently used)}
}
\description{
Predict from SuperRiesz object
}
