% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riesz_superlearner.R
\name{super_riesz}
\alias{super_riesz}
\title{Ensemble estimation of Riesz Representers}
\usage{
super_riesz(
  data,
  library,
  alternatives = list(),
  m = function(alpha, data) alpha(data()),
  folds = 5
)
}
\arguments{
\item{data}{data frame containing observations as originally observed}

\item{library}{vector or list specifying learners to be included in the ensemble}

\item{m}{functional used to define causal parameter of interest for which Riesz Representer is to be estimated}

\item{folds}{number of cross-fitting folds}

\item{data_shifted}{data frame containing observations after treatment intervention has been performed}

\item{conditional_indicator}{matrix indicating which observations are included in conditioning set for causal parameter of interest}
}
\description{
Ensemble estimation of Riesz Representers
}
